Move forks pruning onto a background thread
